Fix test; remove unused structs; remove warnings
