+++ src/lib.rs
@@ -19,6 +19,7 @@ pub mod collection;
 pub mod data;
 pub mod error;
 pub mod schema;
+pub mod value;
 
 mod config;
 mod proto;
