+++ src/schema.rs
@@ -14,53 +14,19 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-use crate::error;
 use crate::error::Result;
+use crate::value::Value;
+use crate::{data::FieldColumn, error};
 use prost::alloc::vec::Vec;
 use std::borrow::Cow;
-use std::collections::HashMap;
-use std::marker::PhantomData;
 use thiserror::Error as ThisError;
 
 use crate::proto::{
     common::KeyValuePair,
-    schema::{
-        self, field_data::Field, scalar_field::Data as ScalarData,
-        vector_field::Data as VectorData, DataType, ScalarField, VectorField,
-    },
+    schema::{self, DataType},
 };
 
 pub use crate::proto::schema::FieldData;
-
-pub trait HasDataType {
-    fn data_type() -> DataType;
-}
-
-macro_rules! impl_has_data_type {
-    ( $($t: ty, $o: expr ),+ ) => {$(
-        impl HasDataType for $t {
-            fn data_type() -> DataType {
-                $o
-            }
-        }
-    )*};
-}
-
-impl_has_data_type! {
-    bool, DataType::Bool,
-    i8, DataType::Int8,
-    i16, DataType::Int16,
-    i32, DataType::Int32,
-    i64, DataType::Int64,
-    f32, DataType::Float,
-    f64, DataType::Double,
-    String, DataType::String,
-    Cow<'_, str>, DataType::String,
-    Vec<f32>, DataType::FloatVector,
-    Vec<u8>, DataType::BinaryVector,
-    Cow<'_, [f32]>, DataType::FloatVector,
-    Cow<'_, [u8]>, DataType::BinaryVector
-}
 
 pub trait Entity {
     const NAME: &'static str;
@@ -550,19 +516,13 @@ pub struct CollectionSchemaBuilder<'a> {
     inner: Vec<FieldSchema<'a>>,
 }
 
-impl<'a> Default for CollectionSchemaBuilder<'a> {
-    fn default() -> Self {
+impl<'a> CollectionSchemaBuilder<'a> {
+    pub fn new(name: Cow<'a, str>, description: Option<Cow<'a, str>>) -> Self {
         Self {
-            name: Cow::Borrowed(""),
-            description: None,
+            name,
+            description,
             inner: Vec::new(),
         }
-    }
-}
-
-impl<'a> CollectionSchemaBuilder<'a> {
-    pub fn new() -> Self {
-        Default::default()
     }
 
     pub fn add_field(&mut self, schema: FieldSchema<'a>) -> &mut Self {
@@ -631,7 +591,7 @@ impl<'a> CollectionSchemaBuilder<'a> {
             return Err(error::Error::from(Error::NoPrimaryKey));
         }
 
-        let this = std::mem::take(self);
+        let this = std::mem::replace(self, CollectionSchemaBuilder::new("".into(), None));
 
         Ok(CollectionSchema {
             fields: this.inner.into(),
@@ -668,530 +628,11 @@ pub enum Error {
     NoSuchKey(String),
 }
 
-pub enum Value<'a> {
-    None,
-    Bool(bool),
-    Int8(i8),
-    Int16(i16),
-    Int32(i32),
-    Long(i64),
-    Float(f32),
-    Double(f64),
-    FloatArray(Cow<'a, [f32]>),
-    Binary(Cow<'a, [u8]>),
-    String(Cow<'a, str>),
-}
-
-macro_rules! impl_from_for_field_data_column {
-    ( $($t: ty, $o: ident ),+ ) => {$(
-        impl From<$t> for Value<'static> {
-            fn from(v: $t) -> Self {
-                Self::$o(v as _)
-            }
-        }
-    )*};
-}
-
-impl_from_for_field_data_column! {
-    bool, Bool,
-    i8,  Int8,
-    i16, Int16,
-    i32, Int32,
-    i64, Long,
-    f32, Float,
-    f64, Double
-}
-
-impl Value<'_> {
-    fn data_type(&self) -> DataType {
-        match self {
-            Value::None => DataType::None,
-            Value::Bool(_) => DataType::Bool,
-            Value::Int8(_) => DataType::Int8,
-            Value::Int16(_) => DataType::Int16,
-            Value::Int32(_) => DataType::Int32,
-            Value::Long(_) => DataType::Int64,
-            Value::Float(_) => DataType::Float,
-            Value::Double(_) => DataType::Double,
-            Value::String(_) => DataType::String,
-            Value::FloatArray(_) => DataType::FloatVector,
-            Value::Binary(_) => DataType::BinaryVector,
-        }
-    }
-}
-
-impl From<String> for Value<'static> {
-    fn from(v: String) -> Self {
-        Self::String(Cow::Owned(v))
-    }
-}
-
-impl<'a> From<&'a str> for Value<'a> {
-    fn from(v: &'a str) -> Self {
-        Self::String(Cow::Borrowed(v))
-    }
-}
-
-impl<'a> From<&'a [u8]> for Value<'a> {
-    fn from(v: &'a [u8]) -> Self {
-        Self::Binary(Cow::Borrowed(v))
-    }
-}
-
-impl From<Vec<u8>> for Value<'static> {
-    fn from(v: Vec<u8>) -> Self {
-        Self::Binary(Cow::Owned(v))
-    }
-}
-
-impl<'a> From<&'a [f32]> for Value<'a> {
-    fn from(v: &'a [f32]) -> Self {
-        Self::FloatArray(Cow::Borrowed(v))
-    }
-}
-
-impl From<Vec<f32>> for Value<'static> {
-    fn from(v: Vec<f32>) -> Self {
-        Self::FloatArray(Cow::Owned(v))
-    }
-}
-
-#[derive(Debug, Clone)]
-pub enum ValueVec {
-    None,
-    Bool(Vec<bool>),
-    Int(Vec<i32>),
-    Long(Vec<i64>),
-    Float(Vec<f32>),
-    Double(Vec<f64>),
-    Binary(Vec<u8>),
-    String(Vec<String>),
-}
-
-macro_rules! impl_from_for_value_vec {
-    ( $($t: ty, $o: ident ),+ ) => {$(
-        impl From<$t> for ValueVec {
-            fn from(v: $t) -> Self {
-                Self::$o(v)
-            }
-        }
-    )*};
-}
-
-impl_from_for_value_vec! {
-    Vec<bool>, Bool,
-    Vec<i32>, Int,
-    Vec<i64>, Long,
-    Vec<String>, String,
-    Vec<u8>, Binary,
-    Vec<f32>, Float,
-    Vec<f64>, Double
-}
-
-impl From<Vec<i8>> for ValueVec {
-    fn from(v: Vec<i8>) -> Self {
-        Self::Int(v.into_iter().map(Into::into).collect())
-    }
-}
-
-impl From<Vec<i16>> for ValueVec {
-    fn from(v: Vec<i16>) -> Self {
-        Self::Int(v.into_iter().map(Into::into).collect())
-    }
-}
-
-impl ValueVec {
-    pub fn new(dtype: DataType) -> Self {
-        match dtype {
-            DataType::None => Self::None,
-            DataType::Bool => Self::Bool(Vec::new()),
-            DataType::Int8 => Self::Int(Vec::new()),
-            DataType::Int16 => Self::Int(Vec::new()),
-            DataType::Int32 => Self::Int(Vec::new()),
-            DataType::Int64 => Self::Long(Vec::new()),
-            DataType::Float => Self::Float(Vec::new()),
-            DataType::Double => Self::Double(Vec::new()),
-            DataType::String => Self::String(Vec::new()),
-            DataType::VarChar => Self::String(Vec::new()),
-            DataType::BinaryVector => Self::Binary(Vec::new()),
-            DataType::FloatVector => Self::Float(Vec::new()),
-        }
-    }
-
-    pub fn check_dtype(&self, dtype: DataType) -> bool {
-        match (self, dtype) {
-            (ValueVec::Binary(..), DataType::BinaryVector)
-            | (ValueVec::Float(..), DataType::FloatVector)
-            | (ValueVec::Float(..), DataType::Float)
-            | (ValueVec::Int(..), DataType::Int8)
-            | (ValueVec::Int(..), DataType::Int16)
-            | (ValueVec::Int(..), DataType::Int32)
-            | (ValueVec::Long(..), DataType::Int64)
-            | (ValueVec::Bool(..), DataType::Bool)
-            | (ValueVec::String(..), DataType::String)
-            | (ValueVec::String(..), DataType::VarChar)
-            | (ValueVec::None, _)
-            | (ValueVec::Double(..), DataType::Double) => true,
-            _ => false,
-        }
-    }
-
-    #[inline]
-    pub fn is_empty(&self) -> bool {
-        self.len() == 0
-    }
-
-    pub fn len(&self) -> usize {
-        match self {
-            ValueVec::None => 0,
-            ValueVec::Bool(v) => v.len(),
-            ValueVec::Int(v) => v.len(),
-            ValueVec::Long(v) => v.len(),
-            ValueVec::Float(v) => v.len(),
-            ValueVec::Double(v) => v.len(),
-            ValueVec::Binary(v) => v.len(),
-            ValueVec::String(v) => v.len(),
-        }
-    }
-
-    pub fn clear(&mut self) {
-        match self {
-            ValueVec::None => (),
-            ValueVec::Bool(v) => v.clear(),
-            ValueVec::Int(v) => v.clear(),
-            ValueVec::Long(v) => v.clear(),
-            ValueVec::Float(v) => v.clear(),
-            ValueVec::Double(v) => v.clear(),
-            ValueVec::Binary(v) => v.clear(),
-            ValueVec::String(v) => v.clear(),
-        }
-    }
-}
-
-impl From<Field> for ValueVec {
-    fn from(f: Field) -> Self {
-        match f {
-            Field::Scalars(s) => match s.data {
-                Some(x) => match x {
-                    ScalarData::BoolData(v) => Self::Bool(v.data),
-                    ScalarData::IntData(v) => Self::Int(v.data),
-                    ScalarData::LongData(v) => Self::Long(v.data),
-                    ScalarData::FloatData(v) => Self::Float(v.data),
-                    ScalarData::DoubleData(v) => Self::Double(v.data),
-                    ScalarData::StringData(v) => Self::String(v.data),
-                    ScalarData::BytesData(v) => unimplemented!(), // Self::Bytes(v.data),
-                },
-                None => Self::None,
-            },
-
-            Field::Vectors(arr) => match arr.data {
-                Some(x) => match x {
-                    VectorData::FloatVector(v) => Self::Float(v.data),
-                    VectorData::BinaryVector(v) => Self::Binary(v),
-                },
-                None => Self::None,
-            },
-        }
-    }
-}
-
-#[derive(Debug, Clone)]
-pub struct FieldColumn<'a> {
-    name: Cow<'a, str>,
-    dtype: DataType,
-    value: ValueVec,
-    dim: i64,
-    max_length: i32,
-}
-
-impl<'a> From<schema::FieldData> for FieldColumn<'a> {
-    fn from(fd: schema::FieldData) -> Self {
-        let (dim, max_length) = fd
-            .field
-            .as_ref()
-            .map(get_dim_max_length)
-            .unwrap_or((None, None));
-
-        let value: ValueVec = fd.field.map(Into::into).unwrap_or(ValueVec::None);
-        let dtype = DataType::from_i32(fd.r#type).unwrap_or(DataType::None);
-
-        FieldColumn {
-            name: Cow::Owned(fd.field_name),
-            dtype,
-            dim: dim.unwrap_or_else(|| match dtype {
-                DataType::None => 0,
-                DataType::Bool
-                | DataType::Int8
-                | DataType::Int16
-                | DataType::Int32
-                | DataType::Int64
-                | DataType::Float
-                | DataType::Double
-                | DataType::String
-                | DataType::VarChar => 1,
-                DataType::BinaryVector => 256,
-                DataType::FloatVector => 128,
-            }),
-            max_length: max_length.unwrap_or(0),
-            value,
-        }
-    }
-}
-
-impl<'a> FieldColumn<'a> {
-    pub fn index(&self, idx: usize) -> Option<Value<'_>> {
-        unimplemented!()
-    }
-
-    fn push(&mut self, val: Value) {
-        match (&mut self.value, val) {
-            (ValueVec::None, Value::None) => (),
-            (ValueVec::Bool(vec), Value::Bool(i)) => vec.push(i),
-            (ValueVec::Int(vec), Value::Int8(i)) => vec.push(i as _),
-            (ValueVec::Int(vec), Value::Int16(i)) => vec.push(i as _),
-            (ValueVec::Int(vec), Value::Int32(i)) => vec.push(i),
-            (ValueVec::Long(vec), Value::Long(i)) => vec.push(i),
-            (ValueVec::Float(vec), Value::Float(i)) => vec.push(i),
-            (ValueVec::Double(vec), Value::Double(i)) => vec.push(i),
-            (ValueVec::String(vec), Value::String(i)) => vec.push(i.to_string()),
-            (ValueVec::Binary(vec), Value::Binary(i)) => vec.extend_from_slice(i.as_ref()),
-            (ValueVec::Float(vec), Value::FloatArray(i)) => vec.extend_from_slice(i.as_ref()),
-            _ => panic!("column type mismatch"),
-        }
-    }
-
-    #[inline]
-    fn len(&self) -> usize {
-        self.value.len()
-    }
-}
-
-fn get_dim_max_length(field: &Field) -> (Option<i64>, Option<i32>) {
-    let dim = match field {
-        Field::Scalars(ScalarField { data: Some(_) }) => 1i64,
-        Field::Vectors(VectorField { dim, .. }) => *dim,
-        _ => 0i64,
-    };
-
-    (Some(dim), None) // no idea how to get max_length
-}
-
-impl<'a> From<FieldColumn<'a>> for schema::FieldData {
-    fn from(this: FieldColumn<'a>) -> schema::FieldData {
-        schema::FieldData {
-            field_name: this.name.to_string(),
-            field_id: 0,
-            r#type: this.dtype as _,
-            field: Some(match this.value {
-                ValueVec::None => Field::Scalars(ScalarField { data: None }),
-                ValueVec::Bool(v) => Field::Scalars(ScalarField {
-                    data: Some(ScalarData::BoolData(schema::BoolArray { data: v })),
-                }),
-                ValueVec::Int(v) => Field::Scalars(ScalarField {
-                    data: Some(ScalarData::IntData(schema::IntArray { data: v })),
-                }),
-                ValueVec::Long(v) => Field::Scalars(ScalarField {
-                    data: Some(ScalarData::LongData(schema::LongArray { data: v })),
-                }),
-                ValueVec::Float(v) => match this.dtype {
-                    DataType::Float => Field::Scalars(ScalarField {
-                        data: Some(ScalarData::FloatData(schema::FloatArray { data: v })),
-                    }),
-                    DataType::FloatVector => Field::Vectors(VectorField {
-                        data: Some(VectorData::FloatVector(schema::FloatArray { data: v })),
-                        dim: this.dim,
-                    }),
-                    _ => unimplemented!(),
-                },
-                ValueVec::Double(v) => Field::Scalars(ScalarField {
-                    data: Some(ScalarData::DoubleData(schema::DoubleArray { data: v })),
-                }),
-                ValueVec::String(v) => Field::Scalars(ScalarField {
-                    data: Some(ScalarData::StringData(schema::StringArray { data: v })),
-                }),
-                ValueVec::Binary(v) => Field::Vectors(VectorField {
-                    data: Some(VectorData::BinaryVector(v)),
-                    dim: this.dim,
-                }),
-            }),
-        }
-    }
-}
-
-pub struct CollectionDataBatch<E> {
-    num_rows: usize,
-    columns: HashMap<Cow<'static, str>, FieldColumn<'static>>,
-    _m: PhantomData<E>,
-}
-
-impl<E: Entity> From<Vec<FieldData>> for CollectionDataBatch<E> {
-    fn from(data: Vec<FieldData>) -> Self {
-        let columns: HashMap<Cow<'static, str>, FieldColumn> = data
-            .into_iter()
-            .map(|mut fld| (std::mem::take(&mut fld.field_name).into(), fld.into()))
-            .collect();
-
-        let schema = E::schema();
-        let primary = schema.primary_column().unwrap();
-        let num_rows = columns.get(&primary.name).unwrap().len();
-
-        Self {
-            num_rows,
-            columns,
-            _m: Default::default(),
-        }
-    }
-}
-
-impl<E> From<CollectionDataBatch<E>> for Vec<FieldData> {
-    fn from(batch: CollectionDataBatch<E>) -> Vec<FieldData> {
-        batch.columns.into_values().map(Into::into).collect()
-    }
-}
-
-impl<E: Entity> Default for CollectionDataBatch<E> {
-    fn default() -> Self {
-        Self::new()
-    }
-}
-
-impl<E> CollectionDataBatch<E> {
-    #[inline]
-    pub fn is_empty(&self) -> bool {
-        self.num_rows == 0
-    }
-
-    #[inline]
-    pub fn len(&self) -> usize {
-        self.num_rows
-    }
-}
-
-impl<E: Entity> CollectionDataBatch<E> {
-    pub fn new() -> Self {
-        let columns: HashMap<_, _> = E::SCHEMA
-            .iter()
-            .map(|x| {
-                (
-                    x.name.clone(),
-                    FieldColumn {
-                        name: x.name.clone(),
-                        dtype: x.dtype,
-                        value: ValueVec::new(x.dtype),
-                        dim: x.dim,
-                        max_length: x.max_length,
-                    },
-                )
-            })
-            .collect();
-
-        Self {
-            num_rows: 0,
-            columns,
-            _m: Default::default(),
-        }
-    }
-
-    #[inline]
-    pub fn add(&mut self, entity: E) {
-        for (schm, val) in entity.iter() {
-            self.columns.get_mut(schm.name.as_ref()).unwrap().push(val);
-        }
-    }
-
-    #[inline]
-    pub fn column(&self, name: &str) -> Option<&FieldColumn> {
-        self.columns.get(name)
-    }
-}
-
-pub struct CollectionDataBatchBuilder<E> {
-    inner: CollectionDataBatch<E>,
-}
-
-impl<E: Entity> CollectionDataBatchBuilder<E> {
-    pub fn new() -> Self {
-        Self {
-            inner: CollectionDataBatch::new(),
-        }
-    }
-
-    pub fn set_column_data<D: Into<ValueVec>>(&mut self, col: &str, data: D) {
-        let data: ValueVec = data.into();
-        let col = self.inner.columns.get_mut(col).unwrap();
-
-        assert!(data.check_dtype(col.dtype));
-
-        col.value = data;
-    }
-
-    pub fn build(&mut self) -> Result<CollectionDataBatch<E>> {
-        Ok(std::mem::take(&mut self.inner))
-    }
-}
-
-pub trait FromField: Sized {
-    fn from_field(field: Field) -> Option<Self>;
-}
-
-macro_rules! impl_from_field {
-    ( $( $t: ty [$($e:tt)*] ),+ ) => {$(
-
-        impl FromField for $t {
-            fn from_field(v: Field) -> Option<Self> {
-                match v {
-                    $($e)*,
-                    _ => None
-                }
-            }
-        }
-    )*};
-}
-
-impl_from_field! {
-    Vec<bool>[Field::Scalars(ScalarField {data: Some(ScalarData::BoolData(schema::BoolArray { data }))}) => Some(data)],
-    Vec<i8>[Field::Scalars(ScalarField {data: Some(ScalarData::IntData(schema::IntArray { data }))}) => Some(data.into_iter().map(|x|x as _).collect())],
-    Vec<i16>[Field::Scalars(ScalarField {data: Some(ScalarData::IntData(schema::IntArray { data }))}) => Some(data.into_iter().map(|x|x as _).collect())],
-    Vec<i32>[Field::Scalars(ScalarField {data: Some(ScalarData::IntData(schema::IntArray { data }))}) => Some(data)],
-    Vec<i64>[Field::Scalars(ScalarField {data: Some(ScalarData::LongData(schema::LongArray { data }))}) => Some(data)],
-    Vec<String>[Field::Scalars(ScalarField {data: Some(ScalarData::StringData(schema::StringArray { data }))}) => Some(data)],
-    Vec<f64>[Field::Scalars(ScalarField {data: Some(ScalarData::DoubleData(schema::DoubleArray { data }))}) => Some(data)],
-    Vec<u8>[Field::Vectors(VectorField {data: Some(VectorData::BinaryVector(data)), ..}) => Some(data)]
-}
-
-impl FromField for Vec<f32> {
-    fn from_field(field: Field) -> Option<Self> {
-        match field {
-            Field::Scalars(ScalarField {
-                data: Some(ScalarData::FloatData(schema::FloatArray { data })),
-            }) => Some(data),
-
-            Field::Vectors(VectorField {
-                data: Some(VectorData::FloatVector(schema::FloatArray { data })),
-                ..
-            }) => Some(data),
-
-            _ => None,
-        }
-    }
-}
-
-pub fn make_field_data<V: Into<ValueVec>>(schm: &FieldSchema, v: V) -> FieldData {
-    let field_column = FieldColumn {
-        name: schm.name.clone(),
-        dtype: schm.dtype,
-        value: v.into(),
-        dim: schm.dim,
-        max_length: schm.max_length,
-    };
-
-    field_column.into()
-}
-
 #[cfg(test)]
 mod tests {
+    use crate::value::Value;
 
-    use super::{FieldSchema, Value};
+    use super::FieldSchema;
 
     struct Test {
         pub id: i64,
