+++ src/value.rs
@@ -1,0 +1,232 @@
+use std::borrow::Cow;
+
+use crate::proto::schema::{
+    field_data::Field, scalar_field::Data as ScalarData, vector_field::Data as VectorData, DataType,
+};
+
+pub enum Value<'a> {
+    None,
+    Bool(bool),
+    Int8(i8),
+    Int16(i16),
+    Int32(i32),
+    Long(i64),
+    Float(f32),
+    Double(f64),
+    FloatArray(Cow<'a, [f32]>),
+    Binary(Cow<'a, [u8]>),
+    String(Cow<'a, str>),
+}
+
+macro_rules! impl_from_for_field_data_column {
+    ( $($t: ty, $o: ident ),+ ) => {$(
+        impl From<$t> for Value<'static> {
+            fn from(v: $t) -> Self {
+                Self::$o(v as _)
+            }
+        }
+    )*};
+}
+
+impl_from_for_field_data_column! {
+    bool, Bool,
+    i8,  Int8,
+    i16, Int16,
+    i32, Int32,
+    i64, Long,
+    f32, Float,
+    f64, Double
+}
+
+impl Value<'_> {
+    pub fn data_type(&self) -> DataType {
+        match self {
+            Value::None => DataType::None,
+            Value::Bool(_) => DataType::Bool,
+            Value::Int8(_) => DataType::Int8,
+            Value::Int16(_) => DataType::Int16,
+            Value::Int32(_) => DataType::Int32,
+            Value::Long(_) => DataType::Int64,
+            Value::Float(_) => DataType::Float,
+            Value::Double(_) => DataType::Double,
+            Value::String(_) => DataType::String,
+            Value::FloatArray(_) => DataType::FloatVector,
+            Value::Binary(_) => DataType::BinaryVector,
+        }
+    }
+}
+
+impl From<String> for Value<'static> {
+    fn from(v: String) -> Self {
+        Self::String(Cow::Owned(v))
+    }
+}
+
+impl<'a> From<&'a str> for Value<'a> {
+    fn from(v: &'a str) -> Self {
+        Self::String(Cow::Borrowed(v))
+    }
+}
+
+impl<'a> From<&'a [u8]> for Value<'a> {
+    fn from(v: &'a [u8]) -> Self {
+        Self::Binary(Cow::Borrowed(v))
+    }
+}
+
+impl From<Vec<u8>> for Value<'static> {
+    fn from(v: Vec<u8>) -> Self {
+        Self::Binary(Cow::Owned(v))
+    }
+}
+
+impl<'a> From<&'a [f32]> for Value<'a> {
+    fn from(v: &'a [f32]) -> Self {
+        Self::FloatArray(Cow::Borrowed(v))
+    }
+}
+
+impl From<Vec<f32>> for Value<'static> {
+    fn from(v: Vec<f32>) -> Self {
+        Self::FloatArray(Cow::Owned(v))
+    }
+}
+
+#[derive(Debug, Clone)]
+pub enum ValueVec {
+    None,
+    Bool(Vec<bool>),
+    Int(Vec<i32>),
+    Long(Vec<i64>),
+    Float(Vec<f32>),
+    Double(Vec<f64>),
+    Binary(Vec<u8>),
+    String(Vec<String>),
+}
+
+macro_rules! impl_from_for_value_vec {
+    ( $($t: ty, $o: ident ),+ ) => {$(
+        impl From<$t> for ValueVec {
+            fn from(v: $t) -> Self {
+                Self::$o(v)
+            }
+        }
+    )*};
+}
+
+impl_from_for_value_vec! {
+    Vec<bool>, Bool,
+    Vec<i32>, Int,
+    Vec<i64>, Long,
+    Vec<String>, String,
+    Vec<u8>, Binary,
+    Vec<f32>, Float,
+    Vec<f64>, Double
+}
+
+impl From<Vec<i8>> for ValueVec {
+    fn from(v: Vec<i8>) -> Self {
+        Self::Int(v.into_iter().map(Into::into).collect())
+    }
+}
+
+impl From<Vec<i16>> for ValueVec {
+    fn from(v: Vec<i16>) -> Self {
+        Self::Int(v.into_iter().map(Into::into).collect())
+    }
+}
+
+impl ValueVec {
+    pub fn new(dtype: DataType) -> Self {
+        match dtype {
+            DataType::None => Self::None,
+            DataType::Bool => Self::Bool(Vec::new()),
+            DataType::Int8 => Self::Int(Vec::new()),
+            DataType::Int16 => Self::Int(Vec::new()),
+            DataType::Int32 => Self::Int(Vec::new()),
+            DataType::Int64 => Self::Long(Vec::new()),
+            DataType::Float => Self::Float(Vec::new()),
+            DataType::Double => Self::Double(Vec::new()),
+            DataType::String => Self::String(Vec::new()),
+            DataType::VarChar => Self::String(Vec::new()),
+            DataType::BinaryVector => Self::Binary(Vec::new()),
+            DataType::FloatVector => Self::Float(Vec::new()),
+        }
+    }
+
+    pub fn check_dtype(&self, dtype: DataType) -> bool {
+        match (self, dtype) {
+            (ValueVec::Binary(..), DataType::BinaryVector)
+            | (ValueVec::Float(..), DataType::FloatVector)
+            | (ValueVec::Float(..), DataType::Float)
+            | (ValueVec::Int(..), DataType::Int8)
+            | (ValueVec::Int(..), DataType::Int16)
+            | (ValueVec::Int(..), DataType::Int32)
+            | (ValueVec::Long(..), DataType::Int64)
+            | (ValueVec::Bool(..), DataType::Bool)
+            | (ValueVec::String(..), DataType::String)
+            | (ValueVec::String(..), DataType::VarChar)
+            | (ValueVec::None, _)
+            | (ValueVec::Double(..), DataType::Double) => true,
+            _ => false,
+        }
+    }
+
+    #[inline]
+    pub fn is_empty(&self) -> bool {
+        self.len() == 0
+    }
+
+    pub fn len(&self) -> usize {
+        match self {
+            ValueVec::None => 0,
+            ValueVec::Bool(v) => v.len(),
+            ValueVec::Int(v) => v.len(),
+            ValueVec::Long(v) => v.len(),
+            ValueVec::Float(v) => v.len(),
+            ValueVec::Double(v) => v.len(),
+            ValueVec::Binary(v) => v.len(),
+            ValueVec::String(v) => v.len(),
+        }
+    }
+
+    pub fn clear(&mut self) {
+        match self {
+            ValueVec::None => (),
+            ValueVec::Bool(v) => v.clear(),
+            ValueVec::Int(v) => v.clear(),
+            ValueVec::Long(v) => v.clear(),
+            ValueVec::Float(v) => v.clear(),
+            ValueVec::Double(v) => v.clear(),
+            ValueVec::Binary(v) => v.clear(),
+            ValueVec::String(v) => v.clear(),
+        }
+    }
+}
+
+impl From<Field> for ValueVec {
+    fn from(f: Field) -> Self {
+        match f {
+            Field::Scalars(s) => match s.data {
+                Some(x) => match x {
+                    ScalarData::BoolData(v) => Self::Bool(v.data),
+                    ScalarData::IntData(v) => Self::Int(v.data),
+                    ScalarData::LongData(v) => Self::Long(v.data),
+                    ScalarData::FloatData(v) => Self::Float(v.data),
+                    ScalarData::DoubleData(v) => Self::Double(v.data),
+                    ScalarData::StringData(v) => Self::String(v.data),
+                    ScalarData::BytesData(_) => unimplemented!(), // Self::Bytes(v.data),
+                },
+                None => Self::None,
+            },
+
+            Field::Vectors(arr) => match arr.data {
+                Some(x) => match x {
+                    VectorData::FloatVector(v) => Self::Float(v.data),
+                    VectorData::BinaryVector(v) => Self::Binary(v),
+                },
+                None => Self::None,
+            },
+        }
+    }
+}
